ics: include definitions for used timezones in ics attachments
The ICS spec requires the TZID to be specified in the icalendar
file. Event though most mail clients can handle standard IANA
timezone strings, not ALL can.

Making use of the tooling built by the chrono-tz team this uses
a new (and for now internal) crate which can generate the timezone
definition for every timezone known to the IANA database.