chore(deps): update dependency minor versions
For a security issue like `RUSTSEC-2024-0336`, we usually update only
the affected package directly. For the minor version update to `rustls`
in `v0.20.9`, the issue is not fixed, so I'm performing an update of all
minor version dependencies which also triggers an update of `rustls` to
`v0.22.4` which fixes the security issue as well.

Closes #94