fix: use connection-level TLS for `smtps://` connections
Before this change, the an attempt was made to open a plain connection
and then upgrade using the STARTTLS command. This does of course not
work if the server already expects TLS encryption on the connection
directly. Therefore the connection should be established with TLS
already on, and without the extra STARTTLS command.

Closes #116